implementation authorization using Gate

+++ app/Http/Controllers/QuestionController.php
@@ -7,6 +7,7 @@ use App\Models\Question;
 use Illuminate\Http\Request;
 
 use DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -44,10 +45,10 @@ class QuestionController extends Controller
     public function store(AskQuestionRequest $request)
     {
         // $request->user()->questions()->create($request->all());
-        $request->user()->questions()->create($request->only('title','body'));
+        $request->user()->questions()->create($request->only('title', 'body'));
 
         // return redirect('/questions');
-        return redirect()->route('questions.index')->with('success',"The Question has been Submitted.");
+        return redirect()->route('questions.index')->with('success', "The Question has been Submitted.");
     }
 
     /**
@@ -70,7 +71,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -85,7 +88,7 @@ class QuestionController extends Controller
     {
         $question->update($request->only('title', 'body'));
 
-        return redirect('/questions')->with('success',"The question has been updated.");
+        return redirect('/questions')->with('success', "The question has been updated.");
     }
 
     /**
@@ -96,8 +99,12 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->delete();
-        
-        return redirect('/questions')->with('success',"The Question has been deleted.");
+
+        return redirect('/questions')->with('success', "The Question has been deleted.");
     }
 }
